updates to enable compiling against ios
